CHE_HX_Asset_Data_Python_Script/requirements.txt
et-xmlfile==1.0.1
jdcal==1.4.1
mysql-connector==2.2.9
mysql-connector-python==8.0.21
numpy==1.19.1
openpyxl==3.0.4
pandas==1.1.0
protobuf==3.12.4
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
xlrd==1.2.0
pytest==6.0.1
